Restructure repo to work with toml file. Fix linting test on commit
