+++ .github/workflows/on-commit.yml
@@ -19,13 +19,12 @@ jobs:
   Lint:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
         with:
           submodules: recursive
-      - uses: s-weigand/setup-conda@v1
+      - uses: actions/setup-python@v5
         with:
-          conda-channels: conda-forge
-          python-version: 3.8
+          python-version: 3.12
       - run: python3 -m venv .venv
       - run: .venv/bin/python -m pip install wheel
       - run: .venv/bin/python setup.py lint
