+++ .gitignore
@@ -26,6 +26,7 @@ share/python-wheels/
 .installed.cfg
 *.egg
 MANIFEST
+resurfemg-dashboard/__version__.py
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
