+++ pyproject.toml
@@ -1,0 +1,53 @@
+[build-system]
+requires = ["setuptools>=64", "setuptools-scm>=8"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "resurfemg-dashboard"
+maintainers = [
+    {name = "Eline Mos-Oppersma", email = "e.oppersma@utwente.nl"},
+    {name = "Rob Warnaar", email = "r.s.p.warnaar@utwente.nl"},
+]
+description = "A GUI package for analysis of respiratory EMG data"
+requires-python = ">=3.9"
+keywords = ["electromyography", "EMG", "respiratory", "Python", "surface EMG",
+"GUI"]
+license = "Apache-2.0"
+license-files = ["LICENSE"]
+classifiers = [
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    'dash-bootstrap-components>=1.2.0',
+    'dash>=2.5.1',
+    'plotly-resampler>=0.7.2.2',
+    'waitress>=2.1.2',
+    'trace_updater',
+    'ipywidgets',
+    'resurfemg==1.0.2'
+]
+readme = "README.md"
+dynamic = ["version"]
+
+[project.urls]
+Homepage = "https://github.com/resurfemg-org"
+Repository = "https://github.com/resurfemg-org/ReSurfEMG-dashboard"
+Issues = "https://github.com/resurfemg-org/ReSurfEMG-dashboard/issues"
+
+[project.optional-dependencies]
+tests = ['pytest', 
+         'pytest-cov',
+         'pycodestyle', 
+         'isort',
+         'wheel',
+         'ipywidgets']
+
+[tool.setuptools.packages.find]
+include = ["resurfemg-dashboard*"]
+
+[tool.setuptools_scm]
+version_file = "resurfemg-dashboard/__version__.py"
