+++ resurfemg-dashboard/static_variables.py
